Added end-to-end tests for ranged for loops (#511)
Turns out you have to explicitly append `f` to the end of floats, so doubles are a more reasonable cross-language floating number. Fine.

Fixes Ruby's "continue" to be "next".

+++ test/integration/ForNumbersStart/number.cs
@@ -1,4 +1,4 @@
 //
-for (float aaa = 5.0; aaa < 10.0; aaa += 1)
+for (double aaa = 5.0; aaa < 10.0; aaa += 1)
 {
 //
